Asm task 08

+++ src_asm/parsing_binary_args.c
@@ -30,7 +30,7 @@ static size_t	save_bin_arg(unsigned char *bin_data, t_arg *arg)
 	return (j);
 }
 
-static void		check_bin_arg(t_line *line, int i, size_t pos,
+static void		check_bin_arg(t_line *line, int i, ssize_t pos,
 										t_file *file)
 {
 	char		*message;
@@ -40,32 +40,31 @@ static void		check_bin_arg(t_line *line, int i, size_t pos,
 	if (!(line->op_data->arg_types[i] & arg.type))
 	{
 		message = make_type_error_message(line, i, file->filename);
-		fill_error(file, NULL, (t_pos)(ssize_t)pos, message);
+		fill_error(file, NULL, (t_pos)pos, message);
 		file->error_data->is_needed_to_free_message = TRUE;
 	}
 	else if ((arg.type == T_REG) && ((char)arg.value < 0))
-		fill_error(file, NULL, (t_pos)(ssize_t)pos,
+		fill_error(file, NULL, (t_pos)pos,
 					"Registry argument value cannot be negative");
 	else if ((arg.type == T_REG) && (arg.value > 99))
-		fill_error(file, NULL, (t_pos)(ssize_t)pos,
+		fill_error(file, NULL, (t_pos)pos,
 					"Registry argument value cannot be larger than 99");
-	
 }
 
 size_t			parse_bin_args(unsigned char *bin_data, t_line *line,
-								size_t pos, t_file *file)
+								ssize_t pos, t_file *file)
 {
 	int			i;
 	size_t		arg_len;
-	size_t		bytes_parsed;
+	ssize_t		bytes_parsed;
 
 	i = 0;
 	bytes_parsed = 0;
 	arg_len = 0;
 	while (i < line->op_data->number_of_args && !(file->error_data))
 	{
-		if ((pos + bytes_parsed + line->args[i].size) > file->code_size)
-			fill_error(file, NULL, (t_pos)(ssize_t)-1,
+		if (((pos + bytes_parsed + line->args[i].size)) > file->code_size)
+			fill_error(file, NULL, (t_pos)(ssize_t)(-1),
 						"Some operations/arguments are incomplete or missing");
 		else
 			arg_len = save_bin_arg(bin_data + bytes_parsed,
