+++ src_asm/parsing_binary_code.c
@@ -12,8 +12,10 @@
 
 #include "asm.h"
 
-static unsigned char	*parse_bin_arg_types(unsigned char *bin_data, t_line *line,
-										size_t pos, t_file *file)
+static unsigned char	*parse_bin_arg_types(unsigned char *bin_data,
+											t_line *line,
+											ssize_t pos,
+											t_file *file)
 {
 	unsigned char		arg_code;
 	unsigned char		*types;
@@ -21,7 +23,7 @@ static unsigned char	*parse_bin_arg_types(unsigned char *bin_data, t_line *line,
 
 	i = 0;
 	if (*bin_data & 3)
-		fill_error(file, NULL, (t_pos)(ssize_t)pos,
+		fill_error(file, NULL, (t_pos)pos,
 					"Expected arguments types byte to end with 2 zero bits");
 	if (!(types = (unsigned char *)ft_memalloc(sizeof(unsigned char) * 3)))
 		exit_with_allocation_error(file->filename);
@@ -35,15 +37,17 @@ static unsigned char	*parse_bin_arg_types(unsigned char *bin_data, t_line *line,
 		else if (arg_code == IND_CODE)
 			types[i] = T_IND;
 		else
-			fill_error(file, NULL, (t_pos)(ssize_t)pos,
+			fill_error(file, NULL, (t_pos)pos,
 						"Incorrect arguments types byte");
 		i++;
 	}
 	return (types);
 }
 
-static size_t			save_bin_arg_types(unsigned char *bin_data, t_line *line,
-										size_t pos, t_file *file)
+static size_t			save_bin_arg_types(unsigned char *bin_data,
+											t_line *line,
+											size_t pos,
+											t_file *file)
 {
 	unsigned char		*types;
 	int					i;
@@ -70,17 +74,17 @@ static size_t			save_bin_arg_types(unsigned char *bin_data, t_line *line,
 	return (line->op_data->has_arg_type_code);
 }
 
-static size_t			parse_bin_op(unsigned char *bin_data, size_t pos,
+static size_t			parse_bin_op(unsigned char *bin_data, ssize_t pos,
 										t_file *file)
 {
 	extern t_op			g_op_tab[OP_NUM];
-	size_t				bytes_parsed;
+	ssize_t				bytes_parsed;
 	t_line				*cur_line;
 
 	bytes_parsed = 1;
 	cur_line = file->last_line;
 	if ((*bin_data > OP_NUM) || (*bin_data == 0))
-		fill_error(file, NULL, (t_pos)(ssize_t)pos, "Incorrect operation code");
+		fill_error(file, NULL, (t_pos)pos, "Incorrect operation code");
 	else if (pos + bytes_parsed >= file->code_size)
 		fill_error(file, NULL, (t_pos)(ssize_t)-1,
 					"Some operations/arguments are incomplete or missing");
@@ -99,6 +103,7 @@ static size_t			parse_bin_op(unsigned char *bin_data, size_t pos,
 /*
 ** [5 spaces][op][space][sym][value][sep][space][next arg]\n
 */
+
 size_t					calc_str_len(t_line *line)
 {
 	size_t				len;
@@ -113,7 +118,7 @@ size_t					calc_str_len(t_line *line)
 		len += ft_strlen(arg.sym) + ft_strlen(arg.str_value);
 		i++;
 		if (i < line->op_data->number_of_args)
-			len += sizeof(char) * 2;		
+			len += sizeof(char) * 2;
 	}
 	len += sizeof(char);
 	return (len);
@@ -121,11 +126,11 @@ size_t					calc_str_len(t_line *line)
 
 void					parse_binary_code(t_file *file)
 {
-	size_t				bytes_parsed;
+	ssize_t				bytes_parsed;
 
 	bytes_parsed = 0;
 	file->first_code_line = file->first_line;
-	while((bytes_parsed < file->code_size) && !(file->error_data))
+	while ((bytes_parsed < (ssize_t)file->code_size) && !(file->error_data))
 	{
 		bytes_parsed += parse_bin_op(file->champ_code + bytes_parsed,
 										bytes_parsed, file);
